moved ModelInt to model folder and added functions
